requirements.txt
pyrogram
tgcrypto
Pillow
pytz
